python/requirements.txt
cycler==0.11.0
et-xmlfile==1.1.0
fonttools==4.37.1
kiwisolver==1.4.4
matplotlib==3.5.3
numpy==1.23.2
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
xlrd==1.2.0
